pyproject.toml
[project]
name = "sourcemanager"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]
packages = [
    { include = "sourcemanager" }
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "coverage (>=7.11.3,<8.0.0)",
    "pytest (>=9.0.0,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]

[tool.poetry.scripts]
flask-app = "src.app:app"